refactor: move background overlay styles to SignInOrJoinFree.js
